logic: fixed warnings, tests and bugs
* fixed clang warnings
* fixed cxxparser tests
* added missing checks for valid location in ASTVisitor callbacks for templates
* removed warning when statusbar view is removed
* made font of activated classes bold
* fixed search box not responding to autocompletion tab or enter press
* fixed search box not updated when active token changes

+++ src/lib/data/parser/cxx/utilityCxx.cpp
@@ -152,7 +152,7 @@ namespace utility
 			{
 				std::string specializedParameterNamePart = "<";
 				const clang::TemplateArgumentList& templateArgumentList = clang::dyn_cast<clang::ClassTemplateSpecializationDecl>(declaration)->getTemplateArgs();
-				for (int i = 0; i < templateArgumentList.size(); i++)
+				for (size_t i = 0; i < templateArgumentList.size(); i++)
 				{
 					DataType datatype = utility::qualTypeToDataType(templateArgumentList.get(i).getAsType());
 					specializedParameterNamePart += datatype.getFullTypeName();
@@ -166,7 +166,7 @@ namespace utility
 		{
 			std::string templateParameterNamePart = "<";
 			clang::TemplateParameterList* parameterList = clang::dyn_cast<clang::TemplateDecl>(declaration)->getTemplateParameters();
-			for (int i = 0; i < parameterList->size(); i++)
+			for (size_t i = 0; i < parameterList->size(); i++)
 			{
 				clang::NamedDecl* namedDecl = parameterList->getParam(i);
